Replace a few more IntFlags with Uint64Flags

+++ beacon-chain/node/node.go
@@ -713,7 +713,7 @@ func (b *BeaconNode) registerRPCService() error {
 	host := b.cliCtx.String(flags.RPCHost.Name)
 	port := b.cliCtx.String(flags.RPCPort.Name)
 	beaconMonitoringHost := b.cliCtx.String(cmd.MonitoringHostFlag.Name)
-	beaconMonitoringPort := b.cliCtx.Int(flags.MonitoringPortFlag.Name)
+	beaconMonitoringPort := b.cliCtx.Uint64(flags.MonitoringPortFlag.Name)
 	cert := b.cliCtx.String(flags.CertFlag.Name)
 	key := b.cliCtx.String(flags.KeyFlag.Name)
 	mockEth1DataVotes := b.cliCtx.Bool(flags.InteropMockEth1DataVotesFlag.Name)
@@ -794,7 +794,7 @@ func (b *BeaconNode) registerPrometheusService(cliCtx *cli.Context) error {
 	additionalHandlers = append(additionalHandlers, prometheus.Handler{Path: "/tree", Handler: c.TreeHandler})
 
 	service := prometheus.NewService(
-		fmt.Sprintf("%s:%d", b.cliCtx.String(cmd.MonitoringHostFlag.Name), b.cliCtx.Int(flags.MonitoringPortFlag.Name)),
+		fmt.Sprintf("%s:%d", b.cliCtx.String(cmd.MonitoringHostFlag.Name), b.cliCtx.Uint64(flags.MonitoringPortFlag.Name)),
 		b.services,
 		additionalHandlers...,
 	)
