+++ validator/rpc/server.go
@@ -33,7 +33,7 @@ type Config struct {
 	ValidatorGatewayHost     string
 	ValidatorGatewayPort     int
 	ValidatorMonitoringHost  string
-	ValidatorMonitoringPort  int
+	ValidatorMonitoringPort  uint64
 	BeaconClientEndpoint     string
 	ClientMaxCallRecvMsgSize int
 	ClientGrpcRetries        uint
@@ -90,7 +90,7 @@ type Server struct {
 	walletInitialized         bool
 	nodeGatewayEndpoint       string
 	validatorMonitoringHost   string
-	validatorMonitoringPort   int
+	validatorMonitoringPort   uint64
 	validatorGatewayHost      string
 	validatorGatewayPort      int
 }
